Remove redundant matrices

+++ include/unoptimized_algo.h
@@ -1,1 +1,1 @@
-int unoptimized_algo(uint8_t[8][8], int[8][8]);
+int unoptimized_algo(int[8][8], int[8][8]);
